+++ apps/demo/src/app/panels/project-panel/project-panel.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { plainToClass } from 'class-transformer-global-storage';
+import { plainToInstance } from 'class-transformer-global-storage';
 import { BehaviorSubject } from 'rxjs';
 import { environment } from '../../../environments/environment';
 import { ProjectPanelStepsEnum } from '../../shared/enums/project-panel-steps.enum';
@@ -9,10 +9,10 @@ import { Project } from '../../shared/models/project';
 export class ProjectPanelService {
   activatedStep$ = new BehaviorSubject(ProjectPanelStepsEnum.Step1);
   project$ = new BehaviorSubject(
-    plainToClass(Project, environment.defaults?.project)
+    plainToInstance(Project, environment.defaults?.project)
   );
   clear() {
-    this.project$.next(plainToClass(Project, environment.defaults?.project));
+    this.project$.next(plainToInstance(Project, environment.defaults?.project));
   }
   store(project: Project) {
     this.project$.next(project);